Set --dirty to .1.dirty in `git describe` call
Without this patch, if you've tagged v1.2.3, and then off that make some
edits, GVB.version would then be 1.2.3.dirty.  The problem is that RubyGems
treats a string as being LESS THAN ZERO, so 1.2.3.dirty < 1.2.3.  In order
to make a dirty tree based off v1.2.3 be considered a newer version than
1.2.3, we need to make the version 1.2.3.1.dirty, which is what
--dirty='.1.dirty' does for us.