Show non-annotated tags as well when listing versions
Currently, creating a tag (such as is done through Github releases) will not be detected by git-version-bump. This is due to the fact that it is a lightweight tag and not an annotated tag. Support for lightweight tags as well as annotated tags would make this much more useful for projects using Github releases.