Fix masked push errors
In case the push of the commit or the tags fails (what is not very
unlikely as it involves network communications), the program ends
without errors, reporting the version was successfully bumped.

The reason is that system() method is invoked without any check of
success or error. This commit takes into account this, assuming
tag_version() method can fail returning false and indicating the error
reason in STDOUT. The main program returns the appropriate error code
when this happens.