Rework the precedence of versioning information sources (fixes #22)
A curious corner case is as follows:

1. an installed gem uses git-version-bump to get its own version.

2. that gem is called in a git repository.

3. the installed gem's version is reported as being that of the code in the git repository (or, if that git repo doesn't use version tags, GVB assplodes).

This is, of course, Very Bad.

Unfortunately, the only solution is to monkey with how GVB decides where to get version information from.
It's always been "use the git repo info if available, otherwise check the calling gem", which is not, in fact, a good approach.
It only works this way because "use the repo, Luke" was how GVB originally worked, and "consult the gemspec" was added later.

This change of approach is a breaking change, and may trigger me to release a v1.0 of GVB.
However, I can't come up with a situation in which you'd *want* a gem to report its version as being from the git repo it's called in, rather than the actual version in the gemspec.
As such, I look forward to the many bug reports that will result from my lack of imagination.